tests/submission/PhilipTaylor/canvas/2d.imageData.object.round.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.imageData.object.round</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.imageData.html">imageData</a>.<a href="index.2d.imageData.object.html">object</a>.round</h1>
<p class="desc">ImageData.data rounds numbers with round-to-zero</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.pixelarray.modify">2d.pixelarray.modify</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData.data rounds numbers with round-to-zero");
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 0.499;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = 0.5;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = 0.501;
_assertSame(imgdata.data[0], 1, "imgdata.data[\""+(0)+"\"]", "1");
imgdata.data[0] = 1.499;
_assertSame(imgdata.data[0], 1, "imgdata.data[\""+(0)+"\"]", "1");
imgdata.data[0] = 1.5;
_assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
imgdata.data[0] = 1.501;
_assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
imgdata.data[0] = 2.5;
_assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
imgdata.data[0] = 3.5;
_assertSame(imgdata.data[0], 4, "imgdata.data[\""+(0)+"\"]", "4");
imgdata.data[0] = 252.5;
_assertSame(imgdata.data[0], 252, "imgdata.data[\""+(0)+"\"]", "252");
imgdata.data[0] = 253.5;
_assertSame(imgdata.data[0], 254, "imgdata.data[\""+(0)+"\"]", "254");
imgdata.data[0] = 254.5;
_assertSame(imgdata.data[0], 254, "imgdata.data[\""+(0)+"\"]", "254");
imgdata.data[0] = 256.5;
_assertSame(imgdata.data[0], 255, "imgdata.data[\""+(0)+"\"]", "255");
imgdata.data[0] = -0.5;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = -1.5;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");


});
</script>

<div id="log"></div>
</body></html>
